requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
anywidget==0.9.18
asttokens==3.0.0
attrs==25.3.0
bbox_visualizer==0.2.0
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
clip-benchmark==1.6.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
datasets==3.5.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
dpcpp-cpp-rt==2025.1.1
einops==0.8.1
executing==2.2.0
fairscale==0.4.13
filelock==3.16.1
filetype==1.2.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.10.0
ftfy==6.3.1
google-auth==2.40.0
google-genai==1.13.0
h11==0.16.0
hf-xet==1.0.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.7
imageio==2.37.0
intel-cmplr-lib-rt==2025.1.1
intel-cmplr-lib-ur==2025.1.1
intel-cmplr-lic-rt==2025.1.1
intel-opencl-rt==2025.1.1
intel-openmp==2025.1.1
intel-pti==0.12.3
intel-sycl-rt==2025.1.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_bbox_widget==0.6.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2025.1.0
mobileclip @ git+https://github.com/apple/ml-mobileclip.git@1140b8d197e4ed7d56b3a92216ded98bb1c2ac87
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.1.1
onemkl-sycl-blas==2025.1.0
onemkl-sycl-dft==2025.1.0
onemkl-sycl-lapack==2025.1.0
onemkl-sycl-rng==2025.1.0
onemkl-sycl-sparse==2025.1.0
onnx==1.16.1
onnx-graphsurgeon==0.5.7
onnxsim==0.4.36
open_clip_torch==2.31.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.15.1
pillow==11.0.0
pillow_heif==0.22.0
platformdirs==4.3.7
polygraphy==0.49.20
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
psygnal==0.12.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
pylabel==0.1.55
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-msssim==1.0.0
pytorch-triton-xpu==3.3.1+gitb0e26b73
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rf100vl==1.0.0
rfdetr==1.1.0
rich==14.0.0
roboflow==1.1.60
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
sentence-transformers==4.1.0
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
supervision==0.25.1
sympy==1.13.3
tbb==2022.1.0
tcmlib==1.3.0
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
tokenizers==0.21.1
torch==2.8.0.dev20250616+xpu
torchaudio==2.8.0.dev20250617+xpu
torchvision==0.23.0.dev20250617+xpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.127
ultralytics-thop==2.0.14
umf==0.10.0
urllib3==2.3.0
wcwidth==0.2.13
webdataset==0.2.111
websockets==15.0.1
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.20.0